feat: add create_grammar_model()
This function creates the Python code for a grammar-aware class that is
compatible with the provided parameter (a Pydantic model class).

"Compatible" in this case means that the grammar-aware class has the same
(annotated) fields as the original class, but with grammar-aware types.
For instance, if the source class has a field called "x" whose type is "str",
the grammar-aware class will have a field called "x" whose type is
"Grammar[str]". We implement support for many common types but this support
is not exhaustive yet.